Copying float branch into default (apart from metadata tunneling, which oduis needs to re-enable n line 278 of simpleprocess.cc).  I copied rtengine into default and patched things up, so this branch should have the rtgui of default and the rtengine of float.

+++ rtengine/PF_correct_RT.cc
@@ -60,13 +60,13 @@ void ImProcFunctions::PF_correct_RT(LabImage * src, LabImage * dst, double radiu
 #endif
 	{
 		AlignedBuffer<double>* buffer = new AlignedBuffer<double> (MAX(src->W,src->H));
-		gaussHorizontal<short> (src->a, tmp1->a, buffer, src->W, src->H, radius, multiThread);
-		gaussHorizontal<short> (src->b, tmp1->b, buffer, src->W, src->H, radius, multiThread);
-		gaussVertical<short>   (tmp1->a, tmp1->a, buffer, src->W, src->H, radius, multiThread);
-		gaussVertical<short>   (tmp1->b, tmp1->b, buffer, src->W, src->H, radius, multiThread);
+		gaussHorizontal<float> (src->a, tmp1->a, buffer, src->W, src->H, radius, multiThread);
+		gaussHorizontal<float> (src->b, tmp1->b, buffer, src->W, src->H, radius, multiThread);
+		gaussVertical<float>   (tmp1->a, tmp1->a, buffer, src->W, src->H, radius, multiThread);
+		gaussVertical<float>   (tmp1->b, tmp1->b, buffer, src->W, src->H, radius, multiThread);
 
-		gaussHorizontal<unsigned short> (src->L, tmp1->L, buffer, src->W, src->H, radius, multiThread);
-		gaussVertical<unsigned short>   (tmp1->L, tmp1->L, buffer, src->W, src->H, radius, multiThread);
+		gaussHorizontal<float> (src->L, tmp1->L, buffer, src->W, src->H, radius, multiThread);
+		gaussVertical<float>   (tmp1->L, tmp1->L, buffer, src->W, src->H, radius, multiThread);
 
 		delete buffer;
 	}
